+++ Blog/urls.py
@@ -18,5 +18,7 @@ from django.urls import path, include
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('', include(('main.urls', 'main'), namespace='main'))
+    path('', include(('main.urls', 'main'), namespace='main')),
+    path('auth/', include(('auth.urls', 'auth'), namespace='auth'))
+
 ]
