+++ auth/urls.py
@@ -1,0 +1,8 @@
+from django.urls import path
+from auth import views
+
+urlpatterns = [
+
+    path("signup", views.Signup, name="signup"),
+    path("login", views.Login, name="login")
+]
