+++ templates/auth/components/login.html
@@ -1,4 +1,6 @@
-{% extends 'main/layouts/base.html' %}
+{% extends 'base.html' %}
+
+
 {% load static %}
 {% block extra_css %}
      <link href="{% static 'main/css/login.css' %}" rel="stylesheet">
