+++ templates/auth/components/signup.html
@@ -1,4 +1,5 @@
-{% extends 'main/layouts/base.html' %}
+{% extends 'base.html' %}
+
 {% load static %}
