+++ templates/main/index.html
@@ -1,16 +1,7 @@
-{% extends 'main/layouts/base.html' %}
-{% block extra_css %} <!--
-TODO somehow make card.css go to base.css and the use that to disable hover for card's read button in card.html
- -->
-{% endblock %}
+{% extends 'base.html' %}
 {% block content %}
 
 
- <!--
-    TODO   If the user has logged in, Do not show the login button.
-    TODO   If the user has logged in, Do not show the sign up button.
-    TODO   If the user has signed up, they should get auto logged in
-    -->
 <section class="jumbotron text-center">
     <div class="container">
         <h1 class="jumbotron-heading">Arnav's Blog</h1>
